+++ script/test/run_parallel_test.py
@@ -41,6 +41,11 @@ def get_config():
         help="Will run in serial.",
     )
     parser.add_argument(
+        "--coverage",
+        action="store_true",
+        help="Execute coverage file.",
+    )
+    parser.add_argument(
         "--keep_cache",
         action="store_true",
         help=(
@@ -271,6 +276,7 @@ async def test_exec(
     test_name=None,
     install_path=None,
     run_in_sandbox=False,
+    restore_db_image_name="erplibre_base",
 ) -> Tuple[str, int]:
     test_result = ""
     test_status = 0
@@ -567,6 +573,8 @@ async def test_exec(
             "./script/db_restore.py",
             "--database",
             unique_database_name,
+            "--image",
+            restore_db_image_name,
             test_name=test_name,
         )
         test_result += res
@@ -576,11 +584,18 @@ async def test_exec(
     if not test_status and lst_init_module_name:
         # Install required module
         str_test = ",".join(lst_init_module_name)
-        script_name = (
-            "./script/addons/install_addons_dev.sh"
-            if tested_module
-            else "./script/addons/install_addons.sh"
-        )
+        if config.coverage:
+            script_name = (
+                "./script/addons/coverage_install_addons_dev.sh"
+                if tested_module
+                else "./script/addons/coverage_install_addons.sh"
+            )
+        else:
+            script_name = (
+                "./script/addons/install_addons_dev.sh"
+                if tested_module
+                else "./script/addons/install_addons.sh"
+            )
         if new_config_path:
             res, status = await run_command(
                 script_name,
@@ -623,10 +638,15 @@ async def test_exec(
     #     destination_path = install_path
 
     if not test_status and tested_module and generated_module:
+        cmd = (
+            "./script/code_generator/coverage_install_and_test_code_generator.sh"
+            if config.coverage
+            else "./script/code_generator/install_and_test_code_generator.sh"
+        )
         # Finally, the test
         if new_config_path:
             res, status = await run_command(
-                "./script/code_generator/install_and_test_code_generator.sh",
+                cmd,
                 unique_database_name,
                 tested_module,
                 test_generated_path,
@@ -636,7 +656,7 @@ async def test_exec(
             )
         else:
             res, status = await run_command(
-                "./script/code_generator/install_and_test_code_generator.sh",
+                cmd,
                 unique_database_name,
                 tested_module,
                 install_path,
@@ -798,6 +818,33 @@ async def run_code_generator_data_test(config) -> Tuple[str, int]:
     return test_result, test_status
 
 
+async def run_code_generator_export_website_attachments_test(
+    config,
+) -> Tuple[str, int]:
+    test_result = ""
+    test_status = 0
+    lst_generated_module = [
+        "demo_website_attachments_data",
+    ]
+    lst_tested_module = [
+        "code_generator_demo_export_website_attachments",
+    ]
+    # Multiple
+    res, status = await test_exec(
+        config,
+        "./addons/TechnoLibre_odoo-code-generator-template",
+        generated_module=",".join(lst_generated_module),
+        tested_module=",".join(lst_tested_module),
+        test_name="code_generator_export_website_attachments_test",
+        run_in_sandbox=True,
+        restore_db_image_name="test_website_attachments",
+    )
+    test_result += res
+    test_status += status
+
+    return test_result, test_status
+
+
 async def run_code_generator_theme_test(config) -> Tuple[str, int]:
     test_result = ""
     test_status = 0
@@ -864,10 +911,12 @@ async def run_code_generator_website_snippet_test(config) -> Tuple[str, int]:
     lst_generated_module = [
         "demo_website_leaflet",
         "demo_website_snippet",
+        "demo_website_multiple_snippet",
     ]
     lst_tested_module = [
         "code_generator_demo_website_leaflet",
         "code_generator_demo_website_snippet",
+        "code_generator_demo_website_multiple_snippet",
     ]
     # Multiple
     res, status = await test_exec(
@@ -1147,6 +1196,7 @@ def run_all_test(config) -> None:
         # Begin run generic test
         # run_code_generator_generic_all_test(config),
         run_code_generator_data_test(config),
+        run_code_generator_export_website_attachments_test(config),
         run_code_generator_theme_test(config),
         run_code_generator_website_snippet_test(config),
         run_code_generator_demo_generic_test(config),